Split cso.rs into cso/ subdirectory (921 -> ~300 lines each)
Split the 921-line cso.rs into four focused files:
- cso/types.rs (281 lines): constants, SewerType, SewerSystemState, CsoEvent
- cso/systems.rs (349 lines): helper functions, update system, plugin
- cso/overflow_tests.rs (309 lines): overflow, pollution, coverage tests
- cso/mod.rs (23 lines): re-exports all public items
